Add cord-platform and seba charts
Adding the combined set of charts for cord-platform and seba.
These have the same content as the upstream cord-platform and seba
opencord repositories, but I'm adding them here for convenience and ease
of tracking, rather than forking the repositories in iecedge.
These charts have been used to create and publish the charts in the
iecedge github helm repository.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>